migrated do-three proxy classes
