+++ src/threecs.ts
@@ -1,5 +1,10 @@
-export * from "./core/components";
+export * from "./core/proxy/Euler"
+export * from "./core/proxy/Quaternion"
+export * from "./core/proxy/Vector3"
+export * from "./core/proxy/Types"
+export * from "./core/proxy/Object3DEntityMixin"
 
+export * from "./core/components";
 export * from "./core/entities";
 
 export * from "./core/ECS";
